+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 25/6/21
+
+### Added
+
+### Changed
+
+- Documentation corrections to better reflect rules intent
+- Change to armour effect so that it is not cumulative
+
+### Removed
+
 ## [1.0.0] - 25/6/21
 
 ### Added